requirements.txt
gitpython==3.1.41
openai==1.12.0
requests==2.31.0